Replace old order rule with new
